\name{benchmarkids}
\alias{benchmarkids}
\title{
The Rsolnp Benchmark Problems Suite problem id's.
}
\description{
Returns the id's of available benchmark in the Rsolnp Benchmark Problems Suite.
}
\usage{
benchmarkids()
}
\value{
A character vector of problem id's.
}
\references{
W.Hock and K.Schittkowski, \emph{Test Examples for Nonlinear Programming Codes}, 
Lecture Notes in Economics and Mathematical Systems. Springer Verlag, 1981.\cr 
Y.Ye, \emph{Interior algorithms for linear, quadratic, and linearly constrained 
non linear programming}, PhD Thesis, Department of EES Stanford University, 
Stanford CA.\cr
}
\author{
Alexios Ghalanos and Stefan Theussl\cr
Y.Ye (original matlab version of solnp)
}
\examples{
benchmarkids()
}
\keyword{optimize}