\name{benchmarklist}
\docType{data}
\alias{benchmarklist}
\title{data: Internal Rsolnp Benchmark Probelm Suite Dataset}
\description{
Internal dataset and results for the Rsolnp Benchmark Problem Suite. Users
should use the function \code{\link{benchmark}} to access a formatted version
of the benchmarks.
}
\usage{data(benchmarklist)}
\format{A list with benchmark solvers and desciptions.}
\keyword{datasets}
