\name{dji30ret}
\docType{data}
\alias{dji30ret}
\title{data: Dow Jones 30 Constituents Closing Value Log Return}
\description{
Dow Jones 30 Constituents closing value log returns from 1987-03-16 to 2009-02-03 from
yahoo finance. Note that AIG was replaced by KFT (Kraft Foods) on September 22, 2008. This
is not reflected in this data set as that would bring the starting date of the data
to 2001.\cr
The dataset is provided for the benchmarking of portfolio optimization problems.
}
\usage{data(dji30ret)}
\format{A data.frame containing 30x5521 observations.}
\source{Yahoo Finance}
\keyword{datasets}
