% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsolnp-package.R
\docType{package}
\name{Rsolnp-package}
\alias{Rsolnp}
\alias{Rsolnp-package}
\title{Rsolnp: General Non-Linear Optimization}
\description{
General Non-linear Optimization Using Augmented Lagrange Multiplier Method.
}
\author{
\strong{Maintainer}: Alexios Galanos \email{alexios@4dscape.com} (\href{https://orcid.org/0009-0000-9308-0457}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Yinyu Ye (Original author of the solnp Matlab code, Stanford University)
}

Other contributors:
\itemize{
  \item Stefan Theussl \email{Stefan.Theussl@R-project.org} (\href{https://orcid.org/0000-0002-6523-4620}{ORCID}) [contributor]
}

}
\keyword{internal}
