\name{Rsolnp-package}
\alias{Rsolnp-package}
\alias{Rsolnp}
\title{The Rsolnp package}
\description{
The Rsolnp package implements Y.Ye's general nonlinear augmented Lagrange multiplier
method solver (SQP based solver).}
\details{
\tabular{ll}{
Package: \tab Rsolnp\cr
Type: \tab Package\cr
Version: \tab 1.0-7\cr
Date: \tab 2010-10-01\cr
License: \tab GPL\cr
LazyLoad: \tab yes\cr
Depends: \tab stats, truncnorm\cr 
Suggests: \tab multicore\cr}
}
\author{
Alexios Ghalanos and Stefan Theussl
}
\references{
Y.Ye, \emph{Interior algorithms for linear, quadratic, and linearly constrained non linear programming},
PhD Thesis, Department of EES Stanford University, Stanford CA.
}